Add Files via Upload.
Python program requires "pip install requests" and "pip install tkcalendar" in order to run it with python.
ARTCC-Publication-Parser will now download the apt.txt file it needs if you tell it you do not have it, and it will download the FAA meta file for the corresponding ARTCC you chose. This program is functional but is improved in many ways.